refactor(loading): 优化loading目录结构
