+++ packages/devui-vue/devui/loading/src/loading-types.ts
@@ -14,7 +14,7 @@ class View {
   top?: string = '50%';
   left?: string = '50%';
 }
-export const componentProps = {
+export const loadingProps = {
   message: String,
   backdrop: Boolean,
   view: {
@@ -28,7 +28,7 @@ export const componentProps = {
   }
 } as const;
 
-export class LoadingProps {
+export class LoadingOptions {
   target?: Element | null;
   message?: string;
   loadingTemplateRef?: VNode['component'];
@@ -38,10 +38,10 @@ export class LoadingProps {
   zIndex?: number;
 }
 
-export type ComponentProps = ExtractPropTypes<typeof componentProps>;
+export type LoadingProps = ExtractPropTypes<typeof loadingProps>;
 
 export interface TargetHTMLElement extends HTMLElement {
   mask?: HTMLElement;
   instance?: VNode['component'];
-  options?: LoadingProps;
+  options?: LoadingOptions;
 }
