+++ packages/devui-vue/devui/tree/src/tree.tsx
@@ -3,7 +3,7 @@ import type { SetupContext } from 'vue';
 import { treeProps, TreeProps, TreeItem, TreeRootType, Nullable } from './tree-types';
 import { CHECK_CONFIG } from  './config';
 import { preCheckTree, deleteNode, getId } from './util';
-import Loading from '../../loading/src/service';
+import Loading from '../../loading/src/loading-service';
 import Checkbox from '../../checkbox/src/checkbox';
 import useToggle from './composables/use-toggle';
 import useMergeNode from './composables/use-merge-node';
